shop/shop/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework.authtoken import views


urlpatterns = [
    path('admin/', admin.site.urls),
    path('', include('main.urls')),
    path('auth/', include('rest_framework.urls')),
    path('token-auth/', views.obtain_auth_token)
]
