requirements.txt
asgiref==3.7.2
Django==4.2.9
django-appconf==1.0.6
django-imagekit==5.0.0
django-js-asset==2.2.0
django-mptt==0.15.0
django-mptt-admin==2.6.1
djangorestframework==3.14.0
pilkit==3.0
pillow==10.2.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
